moved the foodjournal component to a different folder to break it down into smaller components
