+++ src/components/Dashboard/components/FoodJournal/Components/Breakfast.js
@@ -1,0 +1,12 @@
+import React from "react";
+import { ReactComponent as FullHeart } from "../../../../../assets/FullHeart.svg";
+import { ReactComponent as Heart } from "../../../../../assets/HeartOutline.svg";
+
+const Breakfast = (props) => {
+  return (
+    <div>
+    </div>
+  );
+};
+
+export default Breakfast;
